refactor: export test account IDs from miden-objects
