+++ miden-tx/src/compiler/tests.rs
@@ -1,5 +1,6 @@
 use super::{AccountId, ModuleAst, Note, NoteTarget, Operation, ProgramAst, TransactionComplier};
 use miden_objects::{
+    accounts::ACCOUNT_ID_REGULAR_ACCOUNT_IMMUTABLE_CODE_ON_CHAIN,
     assembly::{AssemblyContext, CodeBlock},
     assets::{Asset, FungibleAsset},
     Felt, FieldElement, Word,
@@ -7,8 +8,6 @@ use miden_objects::{
 
 // CONSTANTS
 // ================================================================================================
-
-const ACCOUNT_ID_REGULAR_ACCOUNT_IMMUTABLE_CODE_ON_CHAIN: u64 = 0b0110011011u64 << 54;
 
 // Mast roots of account procedures:
 const ACCT_PROC_1: &'static str =
